Reorganize files for mixtral model (#119)
